# swagger.yaml
openapi: 3.0.0
info:
  title: store API
  contact: {}
  version: "1.0"
servers:
  - url: https://store-api-bil1.onrender.com/
    variables: {}
paths:
  /products/getall:
    get:
      tags:
        - PRODUCTS
      summary: get all products
      operationId: getallproducts
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
  /products:
    get:
      tags:
        - PRODUCTS
      summary: get product
      description: get products based on name, company, featured, also sort products based on price, name, company etc.., add limits to output json, get different pages and also apply numeric filters...
      operationId: getproduct
      parameters:
        - name: featured
          in: query
          description: search for featured products
          style: form
          explode: true
          schema:
            type: boolean
            example: false
        - name: name
          in: query
          description: search for name of the product
          style: form
          explode: true
          schema:
            type: string
            example:
        - name: company
          in: query
          description: search for company of the product
          style: form
          explode: true
          schema:
            type: string
            example:
        - name: sort
          in: query
          description: sort by name, price, company etc..
          style: form
          explode: true
          schema:
            type: string
            example:
        - name: page
          in: query
          description: select different pages
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example:
        - name: limit
          in: query
          description: add limit to output json
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example:
        - name: numericFilter
          in: query
          description: filtering price and ratings
          style: form
          explode: true
          schema:
            type: string
            example: price<30
      responses:
        "200":
          description: OK
          headers:
            X-Powered-By:
              content:
                text/plain:
                  schema:
                    type: string
                  example: Express
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                  example: "1610"
            ETag:
              content:
                text/plain:
                  schema:
                    type: string
                  example: W/"64a-8m3Nd+HXFqGukSnyHFivJ0dO+HM"
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                  example: Wed, 04 Oct 2023 14:12:42 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: keep-alive
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                  example: timeout=5
          content:
            application/json; charset=utf-8:
              schema:
                allOf:
                  - $ref: "#/components/schemas/getproduct"
                  - example:
                      numberOfProducts: 23
                      products:
                        - _id: 651c5487e8d8152b68ceebfd
                          name: albany table
                          price: 309
                        - _id: 651c5487e8d8152b68ceec09
                          name: sofa set
                          price: 129
                        - _id: 651c5487e8d8152b68ceebfe
                          name: armchair
                          price: 125
                        - _id: 651c5487e8d8152b68ceec0c
                          name: vase table
                          price: 120
                        - _id: 651c5487e8d8152b68ceebfc
                          name: albany sectional
                          price: 109
                        - _id: 651c5487e8d8152b68ceec08
                          name: simple chair
                          price: 109
                        - _id: 651c5487e8d8152b68ceec04
                          name: leather sofa
                          price: 99
                        - _id: 651c5487e8d8152b68ceec0b
                          name: utopia sofa
                          price: 79
                        - _id: 651c5487e8d8152b68ceec02
                          name: entertainment center
                          price: 59
                        - _id: 651c5487e8d8152b68ceec00
                          name: dining table
                          price: 42
                        - _id: 651c5487e8d8152b68ceebff
                          name: bar stool
                          price: 40
                        - _id: 651c5487e8d8152b68ceec0f
                          name: wooden desk
                          price: 40
                        - _id: 651c5487e8d8152b68ceec03
                          name: high-back bench
                          price: 39
                        - _id: 651c5487e8d8152b68ceec05
                          name: modern bookshelf
                          price: 31
                        - _id: 651c5487e8d8152b68ceec06
                          name: modern poster
                          price: 30
                        - _id: 651c5487e8d8152b68ceec07
                          name: shelf
                          price: 30
                        - _id: 651c5487e8d8152b68ceebfb
                          name: accent chair
                          price: 25
                        - _id: 651c5487e8d8152b68ceec0d
                          name: wooden bed
                          price: 25
                        - _id: 651c5487e8d8152b68ceec11
                          name: a first wooden table
                          price: 23
                        - _id: 651c5487e8d8152b68ceec01
                          name: emperor bed
                          price: 23
                        - _id: 651c5487e8d8152b68ceec10
                          name: wooden table
                          price: 23
                        - _id: 651c5487e8d8152b68ceec0a
                          name: suede armchair
                          price: 15
                        - _id: 651c5487e8d8152b68ceec0e
                          name: wooden desk
                          price: 15
              example:
                numberOfProducts: 23
                products:
                  - _id: 651c5487e8d8152b68ceebfd
                    name: albany table
                    price: 309
                  - _id: 651c5487e8d8152b68ceec09
                    name: sofa set
                    price: 129
                  - _id: 651c5487e8d8152b68ceebfe
                    name: armchair
                    price: 125
                  - _id: 651c5487e8d8152b68ceec0c
                    name: vase table
                    price: 120
                  - _id: 651c5487e8d8152b68ceebfc
                    name: albany sectional
                    price: 109
                  - _id: 651c5487e8d8152b68ceec08
                    name: simple chair
                    price: 109
                  - _id: 651c5487e8d8152b68ceec04
                    name: leather sofa
                    price: 99
                  - _id: 651c5487e8d8152b68ceec0b
                    name: utopia sofa
                    price: 79
                  - _id: 651c5487e8d8152b68ceec02
                    name: entertainment center
                    price: 59
                  - _id: 651c5487e8d8152b68ceec00
                    name: dining table
                    price: 42
                  - _id: 651c5487e8d8152b68ceebff
                    name: bar stool
                    price: 40
                  - _id: 651c5487e8d8152b68ceec0f
                    name: wooden desk
                    price: 40
                  - _id: 651c5487e8d8152b68ceec03
                    name: high-back bench
                    price: 39
                  - _id: 651c5487e8d8152b68ceec05
                    name: modern bookshelf
                    price: 31
                  - _id: 651c5487e8d8152b68ceec06
                    name: modern poster
                    price: 30
                  - _id: 651c5487e8d8152b68ceec07
                    name: shelf
                    price: 30
                  - _id: 651c5487e8d8152b68ceebfb
                    name: accent chair
                    price: 25
                  - _id: 651c5487e8d8152b68ceec0d
                    name: wooden bed
                    price: 25
                  - _id: 651c5487e8d8152b68ceec11
                    name: a first wooden table
                    price: 23
                  - _id: 651c5487e8d8152b68ceec01
                    name: emperor bed
                    price: 23
                  - _id: 651c5487e8d8152b68ceec10
                    name: wooden table
                    price: 23
                  - _id: 651c5487e8d8152b68ceec0a
                    name: suede armchair
                    price: 15
                  - _id: 651c5487e8d8152b68ceec0e
                    name: wooden desk
                    price: 15
      deprecated: false
      security: []
components:
  schemas:
    getproduct:
      title: getproduct
      required:
        - numberOfProducts
        - products
      type: object
      properties:
        numberOfProducts:
          type: integer
          format: int32
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
          description: ""
      example:
        numberOfProducts: 23
        products:
          - _id: 651c5487e8d8152b68ceebfd
            name: albany table
            price: 309
          - _id: 651c5487e8d8152b68ceec09
            name: sofa set
            price: 129
          - _id: 651c5487e8d8152b68ceebfe
            name: armchair
            price: 125
          - _id: 651c5487e8d8152b68ceec0c
            name: vase table
            price: 120
          - _id: 651c5487e8d8152b68ceebfc
            name: albany sectional
            price: 109
          - _id: 651c5487e8d8152b68ceec08
            name: simple chair
            price: 109
          - _id: 651c5487e8d8152b68ceec04
            name: leather sofa
            price: 99
          - _id: 651c5487e8d8152b68ceec0b
            name: utopia sofa
            price: 79
          - _id: 651c5487e8d8152b68ceec02
            name: entertainment center
            price: 59
          - _id: 651c5487e8d8152b68ceec00
            name: dining table
            price: 42
          - _id: 651c5487e8d8152b68ceebff
            name: bar stool
            price: 40
          - _id: 651c5487e8d8152b68ceec0f
            name: wooden desk
            price: 40
          - _id: 651c5487e8d8152b68ceec03
            name: high-back bench
            price: 39
          - _id: 651c5487e8d8152b68ceec05
            name: modern bookshelf
            price: 31
          - _id: 651c5487e8d8152b68ceec06
            name: modern poster
            price: 30
          - _id: 651c5487e8d8152b68ceec07
            name: shelf
            price: 30
          - _id: 651c5487e8d8152b68ceebfb
            name: accent chair
            price: 25
          - _id: 651c5487e8d8152b68ceec0d
            name: wooden bed
            price: 25
          - _id: 651c5487e8d8152b68ceec11
            name: a first wooden table
            price: 23
          - _id: 651c5487e8d8152b68ceec01
            name: emperor bed
            price: 23
          - _id: 651c5487e8d8152b68ceec10
            name: wooden table
            price: 23
          - _id: 651c5487e8d8152b68ceec0a
            name: suede armchair
            price: 15
          - _id: 651c5487e8d8152b68ceec0e
            name: wooden desk
            price: 15
    Product:
      title: Product
      required:
        - _id
        - name
        - price
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        price:
          type: integer
          format: int32
      example:
        _id: 651c5487e8d8152b68ceebfd
        name: albany table
        price: 309
tags:
  - name: PRODUCTS
    description: ""